platform: Improve fetch32 and fetch64
memcpy (aka ptr::copy_nonoverlapping) is correct to load data from any
alignment (i.e. even a pointer that is not for example correctly aligned
for u64).

The compiler will be responsible for generating the appropriate load
instructions. On x86 it will be an unproblematic fast load.

Relevant blog post
http://fastcompression.blogspot.com/2015/08/accessing-unaligned-memory.html